Evolved the `ConnectionOptions` into a connection string extractor
The `ConnectionOptions` class can extract the connection string items out and set them as properties.  Defined a rule list instead of a massive switch statement.  Checked unit tests, and tweaked on to cover a bug I was introducing as I went.  That bug is now fixed and the unit tests all pass again.